// src/app/models/enums.ts
export enum BlockComponentNames {
  ABOUTUSINFO = 'aboutusinfo',
  BENEFITSINFO = 'benefitsinfo',
  CLIENTREVIEWS = 'clientreviews',
  CONTACTUS = 'contactus',
  GREETINGWITHIMAGE = 'greetingwithimage',
  GREETINGWITHTEXT = 'greetingwithtext',
  HOWITWORKINFO = 'howitworkinfo',
  INDUSTRIESINFO = 'industriesinfo',
  MISSIONINFOWITHONEIMAGE = 'missioninfowithoneimage',
  MISSIONINFOWITHTWOCOLUMN = 'missioninfowithtwocolumn',
  OURPROJECTSINFO = 'ourprojectsinfo',
  PROVIDEINFO = 'provideinfo',
  PROVIDEWITHBLOCKSINFO = 'providewithblocksinfo',
  SHAREDETAILSINFO = 'sharedetailsinfo',
  TECHNOLOGIESINFO = 'technologiesinfo',
  TRANSPARENTPROCESSINFO = 'transparentprocessinfo',
  WHYDELEGATEINFO = 'whydelegateinfo',
  WHYDOUSEINFO = 'whydouseinfo',
  CASEHEADER = 'caseheader',
  CHALLENGETOSOLUTION = 'challengetosolution',
  PROCESSINFO = 'processinfo',
  SKETCHES = 'sketches',
  WIREFRAMES = 'wireframes',
  TYPOGRAPHYINFO = 'typographyinfo',
  CASEICONSET = 'caseiconset',
  CASEROADMAP = 'caseroadmap',
  CASERESULT = 'caseresult',
  PROCESSINFOWITHDESCRIPTION = 'processinfowithdescription',
  SKETCHESWIREFRAMES = 'sketcheswireframes',
  WHATIS = 'whatis',
  CONTACTFORM = 'contactform'
}